requirements/development.txt
-r base.txt

fake-factory==0.7.4
factory-boy==2.7.0
mock==1.0.1
django-debug-toolbar
flake8==3.6.0
